Update base for Update on "[NOT FOR lAND] Temp commit to allow tracing of on device quant"
methods via bundled inputs
g

Summary:

Test Plan:

Reviewers:

Subscribers:

Tasks:

Tags:

Differential Revision: [D39062963](https://our.internmc.facebook.com/intern/diff/D39062963)

[ghstack-poisoned]

+++ torch/csrc/jit/mobile/module.cpp
@@ -62,7 +62,7 @@ bool Module::compareMethodSchemas(
 }
 
 void Module::unsafeRemoveMethod(const std::string& basename) {
-  size_t i = 0;
+  int64_t i = 0;
   for (; i < cu_->methods().size(); ++i) {
     if ((cu_->methods()[i])->name() == basename) {
       break;
